+++ rotors_joy_interface/CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog for package rotors_joy_interface
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+6.0.2 (2019-12-28)
+------------------
+* Add joystick interface for the Crazyflie
+* Contributors: Giuseppe Silano
+
 6.0.1 (2019-12-28)
 ------------------